refactor: move package of MT module
