+++ machinetranslators/apertium/src/main/java/org/omegat/machinetranslators/apertium/ApertiumTranslate.java
@@ -26,7 +26,7 @@
  *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package org.omegat.core.machinetranslators;
+package org.omegat.machinetranslators.apertium;
 
 import java.awt.Window;
 import java.awt.event.ItemEvent;
@@ -35,6 +35,7 @@ import java.io.IOException;
 import java.net.URL;
 import java.net.URLEncoder;
 import java.util.Locale;
+import java.util.ResourceBundle;
 
 import javax.swing.JCheckBox;
 import javax.swing.event.DocumentEvent;
@@ -43,12 +44,15 @@ import javax.swing.event.DocumentListener;
 import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import tokyo.northside.logging.ILogger;
+import tokyo.northside.logging.LoggerFactory;
 
 import org.omegat.core.Core;
+import org.omegat.core.machinetranslators.BaseCachedTranslate;
+import org.omegat.core.machinetranslators.MachineTranslateError;
 import org.omegat.gui.exttrans.MTConfigDialog;
 import org.omegat.util.HttpConnectionUtils;
 import org.omegat.util.Language;
-import org.omegat.util.Log;
 import org.omegat.util.OStrings;
 import org.omegat.util.Preferences;
 import org.omegat.util.StringUtil;
@@ -60,7 +64,6 @@ import org.omegat.util.StringUtil;
 public class ApertiumTranslate extends BaseCachedTranslate {
 
     private static final int HTTP_OK = 200;
-
     protected static final String PROPERTY_APERTIUM_MARKUNKNOWN = "apertium.server.markunknown";
     protected static final String PROPERTY_APERTIUM_SERVER_CUSTOM = "apertium.server.custom";
     protected static final String PROPERTY_APERTIUM_SERVER_URL = "apertium.server.url";
@@ -69,25 +72,23 @@ public class ApertiumTranslate extends BaseCachedTranslate {
     protected static final String APERTIUM_SERVER_URL_FORMAT = "%s/translate?q=%s&markUnknown=%s&langpair=%s|%s&key=%s";
     // Specific OmegaT key
     protected static final String APERTIUM_SERVER_KEY_DEFAULT = "bwuxb5jS+VwSJ8mLz1qMfmMrDGA";
-    private ObjectMapper mapper = new ObjectMapper();
+    private final ObjectMapper mapper = new ObjectMapper();
+    private final ResourceBundle bundle;
+    private final ILogger logger;
 
     /**
      * Register plugins into OmegaT.
      */
     public static void loadPlugins() {
         Core.registerMachineTranslationClass(ApertiumTranslate.class);
-/*
-        Core.registerMachineTranslationClass(BelazarTranslate.class);
-        Core.registerMachineTranslationClass(DeepLTranslate.class);
-        Core.registerMachineTranslationClass(Google2Translate.class);
-        Core.registerMachineTranslationClass(IBMWatsonTranslate.class);
-        Core.registerMachineTranslationClass(MyMemoryHumanTranslate.class);
-        Core.registerMachineTranslationClass(MyMemoryMachineTranslate.class);
-        Core.registerMachineTranslationClass(YandexCloudTranslate.class);
- */
     }
 
     public static void unloadPlugins() {
+    }
+
+    public ApertiumTranslate() {
+        bundle = ResourceBundle.getBundle("Bundle");
+        logger = LoggerFactory.getLogger(ApertiumTranslate.class);
     }
 
     @Override
@@ -101,7 +102,7 @@ public class ApertiumTranslate extends BaseCachedTranslate {
      * @return engine name.
      */
     public String getName() {
-        return OStrings.getString("MT_ENGINE_APERTIUM");
+        return bundle.getString("MT_ENGINE_APERTIUM");
     }
 
     /**
@@ -148,8 +149,8 @@ public class ApertiumTranslate extends BaseCachedTranslate {
         try {
             v = HttpConnectionUtils.getURL(new URL(url));
         } catch (IOException e) {
-            Log.logErrorRB(e, "APERTIUM_CUSTOM_SERVER_NOTFOUND");
-            throw new Exception(OStrings.getString("APERTIUM_CUSTOM_SERVER_NOTFOUND"));
+            logger.atError().setCause(e).setMessageRB("APERTIUM_CUSTOM_SERVER_NOTFOUND").log();
+            throw new Exception(bundle.getString("APERTIUM_CUSTOM_SERVER_NOTFOUND"));
         }
 
         return getJsonResults(v);
@@ -182,8 +183,8 @@ public class ApertiumTranslate extends BaseCachedTranslate {
             String details = rootNode.get("responseDetails").asText();
             throw new MachineTranslateError(StringUtil.format(OStrings.getString("APERTIUM_ERROR"), code, details));
         } catch (JsonParseException e) {
-            Log.logErrorRB(e, "MT_JSON_ERROR");
-            throw new MachineTranslateError(OStrings.getString("MT_JSON_ERROR"));
+            logger.atError().setCause(e).setMessageRB("MT_JSON_ERROR");
+            throw new MachineTranslateError(bundle.getString("MT_JSON_ERROR"));
         }
     }
