+++ machinetranslators/apertium/src/main/resources/org/omegat/machinetranslators/apertium/Bundle.properties
@@ -22,11 +22,12 @@
 #  You should have received a copy of the GNU General Public License
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
+MT_ENGINE_APERTIUM=Apertium
 
-License= GNU Public License version 3 or later
-Author=OmegaT team
-Name=Apertium Translate connector
-Version=6.0.0
-Category=machinetranslator
-Class=org.omegat.machinetranslators.apertium.ApertiumTranslate
-Description==Apertium Translate
+APERTIUM_ERROR=Error {0}: {1}
+APERTIUM_CUSTOM_SERVER_LABEL=Connect to custom server instead of apertium.org
+APERTIUM_CUSTOM_SERVER_URL_LABEL=Custom server URL
+APERTIUM_CUSTOM_SERVER_KEY_LABEL=API key (optional)
+APERTIUM_CUSTOM_SERVER_NOTFOUND=Unable to connect to server. Ensure the server URL is correct.
+APERTIUM_CUSTOM_SERVER_INVALID=The server is not a valid Apertium server. Ensure the URL is correct.
+APERTIUM_MARKUNKNOWN_LABEL=Mark unknown words
