+++ machinetranslators/apertium/src/main/resources/org/omegat/machinetranslators/apertium/Bundle_be.properties
@@ -1,0 +1,33 @@
+#
+#  OmegaT - Computer Assisted Translation (CAT) tool
+#           with fuzzy matching, translation memory, keyword search,
+#           glossaries, and translation leveraging into updated projects.
+#
+#  Copyright (C) 2023 Hiroshi Miura
+#                Home page: https://www.omegat.org/
+#                Support center: https://omegat.org/support
+#
+#  This file is part of OmegaT.
+#
+#  OmegaT is free software: you can redistribute it and/or modify
+#  it under the terms of the GNU General Public License as published by
+#  the Free Software Foundation, either version 3 of the License, or
+#  (at your option) any later version.
+#
+#  OmegaT is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
+#
+#  You should have received a copy of the GNU General Public License
+#  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+#
+MT_ENGINE_APERTIUM=Apertium
+
+APERTIUM_ERROR=\u041F\u0430\u043C\u044B\u043B\u043A\u0430 {0}: {1}
+APERTIUM_CUSTOM_SERVER_LABEL=\u0417\u043B\u0443\u0447\u044B\u0446\u0446\u0430 \u0437 \u0441\u044D\u0440\u0432\u044D\u0440\u0430\u043C \u043A\u0430\u0440\u044B\u0441\u0442\u0430\u043B\u044C\u043D\u0456\u043A\u0430 \u0437\u0430\u043C\u0435\u0441\u0442 apertium.org
+APERTIUM_CUSTOM_SERVER_URL_LABEL=URL-\u0430\u0434\u0440\u0430\u0441 \u0441\u044D\u0440\u0432\u044D\u0440\u0430 \u043A\u0430\u0440\u044B\u0441\u0442\u0430\u043B\u044C\u043D\u0456\u043A\u0430
+APERTIUM_CUSTOM_SERVER_KEY_LABEL=\u041A\u043B\u044E\u0447 API (\u043A\u0430\u043B\u0456 \u043F\u0430\u0442\u0440\u044D\u0431\u043D\u044B)
+APERTIUM_CUSTOM_SERVER_NOTFOUND=\u041D\u0435\u043C\u0430\u0433\u0447\u044B\u043C\u0430 \u0437\u043B\u0443\u0447\u044B\u0446\u0446\u0430 \u0437 \u0441\u044D\u0440\u0432\u044D\u0440\u0430\u043C. \u0421\u043F\u0440\u0430\u045E\u0434\u0437\u0435\u0446\u0435 URL-\u0430\u0434\u0440\u0430\u0441 \u0441\u044D\u0440\u0432\u044D\u0440\u0430.
+APERTIUM_CUSTOM_SERVER_INVALID=\u0413\u044D\u0442\u044B \u0441\u044D\u0440\u0432\u044D\u0440 \u2014 \u043D\u0435 \u0441\u0430\u043F\u0440\u0430\u045E\u0434\u043D\u044B \u0441\u044D\u0440\u0432\u044D\u0440 Apertium. \u0421\u043F\u0440\u0430\u045E\u0434\u0437\u0435\u0446\u0435 URL-\u0430\u0434\u0440\u0430\u0441.
+APERTIUM_MARKUNKNOWN_LABEL=\u041F\u0430\u0437\u043D\u0430\u0447\u0430\u0446\u044C \u043D\u0435\u0432\u044F\u0434\u043E\u043C\u044B\u044F \u0441\u043B\u043E\u0432\u044B
