+++ machinetranslators/apertium/src/main/resources/org/omegat/machinetranslators/apertium/Bundle_pt.properties
@@ -23,10 +23,12 @@
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
 
-License= GNU Public License version 3 or later
-Author=OmegaT team
-Name=Apertium Translate connector
-Version=6.0.0
-Category=machinetranslator
-Class=org.omegat.machinetranslators.apertium.ApertiumTranslate
-Description==Apertium Translate
+MT_ENGINE_APERTIUM=Apertium
+
+APERTIUM_ERROR=Erro {0}: {1}
+APERTIUM_CUSTOM_SERVER_LABEL=Ligar a servidor personalizado em vez do apertium.org
+APERTIUM_CUSTOM_SERVER_URL_LABEL=URL do servidor personalizado
+APERTIUM_CUSTOM_SERVER_KEY_LABEL=Chave API (opcional)
+APERTIUM_CUSTOM_SERVER_NOTFOUND=Imposs\u00EDvel ligar ao servidor. Assegure-se que o URL est\u00E1 correcto.
+APERTIUM_CUSTOM_SERVER_INVALID=O servidor n\u00E3o \u00E9 um servidor Apertium v\u00E1lido. Assegure-se de que o URL est\u00E1 correcto.
+APERTIUM_MARKUNKNOWN_LABEL=Marcar palavras desconhecidas
