+++ machinetranslators/apertium/src/main/resources/org/omegat/machinetranslators/apertium/Bundle_pt_BR.properties
@@ -23,10 +23,3 @@
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
 
-License= GNU Public License version 3 or later
-Author=OmegaT team
-Name=Apertium Translate connector
-Version=6.0.0
-Category=machinetranslator
-Class=org.omegat.machinetranslators.apertium.ApertiumTranslate
-Description==Apertium Translate
