+++ machinetranslators/apertium/src/main/resources/org/omegat/machinetranslators/apertium/Bundle_ru.properties
@@ -1,0 +1,32 @@
+#
+#  OmegaT - Computer Assisted Translation (CAT) tool
+#           with fuzzy matching, translation memory, keyword search,
+#           glossaries, and translation leveraging into updated projects.
+#
+#  Copyright (C) 2023 Hiroshi Miura
+#                Home page: https://www.omegat.org/
+#                Support center: https://omegat.org/support
+#
+#  This file is part of OmegaT.
+#
+#  OmegaT is free software: you can redistribute it and/or modify
+#  it under the terms of the GNU General Public License as published by
+#  the Free Software Foundation, either version 3 of the License, or
+#  (at your option) any later version.
+#
+#  OmegaT is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
+#
+#  You should have received a copy of the GNU General Public License
+#  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+#
+MT_ENGINE_APERTIUM=Apertium
+APERTIUM_ERROR=\u041E\u0448\u0438\u0431\u043A\u0430 {0}: {1}
+APERTIUM_CUSTOM_SERVER_LABEL=\u041F\u043E\u0434\u043A\u043B\u044E\u0447\u0430\u0442\u044C\u0441\u044F \u043A \u0447\u0430\u0441\u0442\u043D\u043E\u043C\u0443 \u0441\u0435\u0440\u0432\u0435\u0440\u0443 \u0432\u043C\u0435\u0441\u0442\u043E apertium.org
+APERTIUM_CUSTOM_SERVER_URL_LABEL=URL \u0447\u0430\u0441\u0442\u043D\u043E\u0433\u043E \u0441\u0435\u0440\u0432\u0435\u0440\u0430
+APERTIUM_CUSTOM_SERVER_KEY_LABEL=\u041A\u043B\u044E\u0447 API (\u043E\u043F\u0446\u0438\u043E\u043D\u0430\u043B\u044C\u043D\u043E)
+APERTIUM_CUSTOM_SERVER_NOTFOUND=\u041D\u0435 \u0443\u0434\u0430\u043B\u043E\u0441\u044C \u043F\u043E\u0434\u043A\u043B\u044E\u0447\u0438\u0442\u044C\u0441\u044F \u043A \u0441\u0435\u0440\u0432\u0435\u0440\u0443. \u0423\u0431\u0435\u0434\u0438\u0442\u0435\u0441\u044C, \u0447\u0442\u043E URL \u0441\u0435\u0440\u0432\u0435\u0440\u0430 \u0443\u043A\u0430\u0437\u0430\u043D \u0432\u0435\u0440\u043D\u043E.
+APERTIUM_CUSTOM_SERVER_INVALID=\u041D\u0430 \u0441\u0435\u0440\u0432\u0435\u0440\u0435 \u043D\u0435 \u0440\u0430\u0431\u043E\u0442\u0430\u0435\u0442 Apertium. \u0423\u0431\u0435\u0434\u0438\u0442\u0435\u0441\u044C, \u0447\u0442\u043E URL \u0443\u043A\u0430\u0437\u0430\u043D \u0432\u0435\u0440\u043D\u043E.
+APERTIUM_MARKUNKNOWN_LABEL=\u041F\u043E\u043C\u0435\u0447\u0430\u0442\u044C \u043D\u0435\u0438\u0437\u0432\u0435\u0441\u0442\u043D\u044B\u0435 \u0441\u043B\u043E\u0432\u0430
