+++ machinetranslators/apertium/src/main/resources/org/omegat/machinetranslators/apertium/Bundle_sv.properties
@@ -23,10 +23,5 @@
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
 
-License= GNU Public License version 3 or later
-Author=OmegaT team
-Name=Apertium Translate connector
-Version=6.0.0
-Category=machinetranslator
-Class=org.omegat.machinetranslators.apertium.ApertiumTranslate
-Description==Apertium Translate
+MT_ENGINE_APERTIUM=Apertium
+APERTIUM_ERROR=Fel {0}: {1}
