+++ machinetranslators/apertium/src/main/resources/org/omegat/machinetranslators/apertium/Bundle_uk.properties
@@ -1,0 +1,34 @@
+#
+#  OmegaT - Computer Assisted Translation (CAT) tool
+#           with fuzzy matching, translation memory, keyword search,
+#           glossaries, and translation leveraging into updated projects.
+#
+#  Copyright (C) 2023 Hiroshi Miura
+#                Home page: https://www.omegat.org/
+#                Support center: https://omegat.org/support
+#
+#  This file is part of OmegaT.
+#
+#  OmegaT is free software: you can redistribute it and/or modify
+#  it under the terms of the GNU General Public License as published by
+#  the Free Software Foundation, either version 3 of the License, or
+#  (at your option) any later version.
+#
+#  OmegaT is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
+#
+#  You should have received a copy of the GNU General Public License
+#  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+#
+
+MT_ENGINE_APERTIUM=Apertium
+
+APERTIUM_ERROR=\u041F\u043E\u043C\u0438\u043B\u043A\u0430 {0}: {1}
+APERTIUM_CUSTOM_SERVER_LABEL=\u041F\u0456\u0434\u2019\u0454\u0434\u043D\u0430\u0442\u0438\u0441\u044F \u0434\u043E \u0456\u043D\u0448\u043E\u0433\u043E \u0441\u0435\u0440\u0432\u0435\u0440\u0430 \u0437\u0430\u043C\u0456\u0441\u0442\u044C apertium.org
+APERTIUM_CUSTOM_SERVER_URL_LABEL=URL-\u0430\u0434\u0440\u0435\u0441\u0430 \u0456\u043D\u0448\u043E\u0433\u043E \u0441\u0435\u0440\u0432\u0435\u0440\u0430:
+APERTIUM_CUSTOM_SERVER_KEY_LABEL=\u041A\u043B\u044E\u0447 API (\u043D\u0435\u043E\u0431\u043E\u0432\u2019\u044F\u0437\u043A\u043E\u0432\u043E):
+APERTIUM_CUSTOM_SERVER_NOTFOUND=\u041D\u0435 \u0432\u0434\u0430\u043B\u043E\u0441\u044F \u043F\u0456\u0434\u2019\u0454\u0434\u043D\u0430\u0442\u0438\u0441\u044F \u0434\u043E \u0441\u0435\u0440\u0432\u0435\u0440\u0430. \u041F\u0435\u0440\u0435\u0432\u0456\u0440\u0442\u0435 \u043F\u0440\u0430\u0432\u0438\u043B\u044C\u043D\u0456\u0441\u0442\u044C \u0437\u0430\u0437\u043D\u0430\u0447\u0435\u043D\u043E\u0457 URL-\u0430\u0434\u0440\u0435\u0441\u0438.
+APERTIUM_CUSTOM_SERVER_INVALID=\u0423\u043A\u0430\u0437\u0430\u043D\u0438\u0439 \u0441\u0435\u0440\u0432\u0435\u0440 \u043D\u0435 \u0454 \u043D\u0430\u043B\u0435\u0436\u043D\u0438\u043C \u0441\u0435\u0440\u0432\u0435\u0440\u043E\u043C Apertium. \u041F\u0435\u0440\u0435\u0432\u0456\u0440\u0442\u0435 \u043F\u0440\u0430\u0432\u0438\u043B\u044C\u043D\u0456\u0441\u0442\u044C \u0437\u0430\u0437\u043D\u0430\u0447\u0435\u043D\u043E\u0457 URL-\u0430\u0434\u0440\u0435\u0441\u0438.
+APERTIUM_MARKUNKNOWN_LABEL=\u041F\u043E\u0437\u043D\u0430\u0447\u0430\u0442\u0438 \u043D\u0435\u0432\u0456\u0434\u043E\u043C\u0456 \u0441\u043B\u043E\u0432\u0430
