+++ machinetranslators/apertium/src/main/resources/org/omegat/machinetranslators/apertium/Bundle_zh_CN.properties
@@ -23,10 +23,12 @@
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
 
-License= GNU Public License version 3 or later
-Author=OmegaT team
-Name=Apertium Translate connector
-Version=6.0.0
-Category=machinetranslator
-Class=org.omegat.machinetranslators.apertium.ApertiumTranslate
-Description==Apertium Translate
+MT_ENGINE_APERTIUM=Apertium
+
+APERTIUM_ERROR=\u9519\u8BEF {0}: {1}
+APERTIUM_CUSTOM_SERVER_LABEL=\u8FDE\u63A5\u5230\u81EA\u5B9A\u4E49\u670D\u52A1\u5668\u800C\u975E apertium.org
+APERTIUM_CUSTOM_SERVER_URL_LABEL=\u81EA\u5B9A\u4E49\u670D\u52A1\u5668 URL
+APERTIUM_CUSTOM_SERVER_KEY_LABEL=API \u5BC6\u94A5\uFF08\u53EF\u9009\uFF09
+APERTIUM_CUSTOM_SERVER_NOTFOUND=\u65E0\u6CD5\u8FDE\u63A5\u5230\u670D\u52A1\u5668\u3002\u8BF7\u786E\u4FDD\u670D\u52A1\u5668 URL \u6B63\u786E\u3002
+APERTIUM_CUSTOM_SERVER_INVALID=\u6B64\u670D\u52A1\u5668\u4E0D\u662F\u6709\u6548\u7684 Apertium \u670D\u52A1\u5668\u3002\u8BF7\u786E\u4FDD URL \u6B63\u786E\u3002
+APERTIUM_MARKUNKNOWN_LABEL=\u6807\u8BB0\u672A\u77E5\u5355\u8BCD
