+++ machinetranslators/apertium/src/main/resources/org/omegat/machinetranslators/apertium/Bundle_zh_TW.properties
@@ -22,11 +22,10 @@
 #  You should have received a copy of the GNU General Public License
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
-
-License= GNU Public License version 3 or later
-Author=OmegaT team
-Name=Apertium Translate connector
-Version=6.0.0
-Category=machinetranslator
-Class=org.omegat.machinetranslators.apertium.ApertiumTranslate
-Description==Apertium Translate
+APERTIUM_CUSTOM_SERVER_INVALID=
+APERTIUM_CUSTOM_SERVER_KEY_LABEL=
+APERTIUM_CUSTOM_SERVER_LABEL=
+APERTIUM_CUSTOM_SERVER_NOTFOUND=
+APERTIUM_ERROR=
+APERTIUM_MARKUNKNOWN_LABEL=
+MT_ENGINE_APERTIUM=Apertium
